% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_moments.R
\name{trait_moments}
\alias{trait_moments}
\title{trait_moments}
\usage{
trait_moments(communities, traits, n_species = 4, abundance = 80)
}
\arguments{
\item{communities}{The species-plot-matrix as data frame. Column names must be species IDs and row names community IDs. It is not mandatory to provide relative abundances, but abundances should be on a nominal scale.}

\item{traits}{The species-trait-matrix as data frame. Column names must be trait IDs and row names species IDs. The species IDs must be consistent with those from the species-plot-matrix. Trait values have to be numeric.}

\item{n_species}{Number of predominant species for which trait information must be present. If the criterion is not met, NA is returned for the trait-plot-combination. Default is 4.}

\item{abundance}{Cumulative relative abundance for which trait information must be present. If the criterion is not met, NA is returned for the trait-plot-combination. Default is 80\%.}
}
\value{
Mean (CWM), variance, skewness and kurtosis of all trait distributions for each trait and each community.
}
\description{
trait_moments calculates all moments (mean, variance, skewness and kurtosis) of trait distribution across large community and traits datasets. The calculation is performed according to the equations number 1 to 4 from Le Bagousse-Pinguet et al. (2017):
}
\details{
[
\small
Mean_{j} = \sum_{i}^{n} p_i T_i
\tag{1}
]

[
\small
Variance_{j} = \sum_{i}^{n} p_i (T_i - Mean_j)^2
\tag{2}
]

[
\small
Skewness_{j} = \sum_{i}^{n} \frac{p_i (T_i - Mean_j)^3}{Variance_j^{\frac{3}{2}}}
\tag{3}
]

[
\small
Kurtosis_{j} = \sum_{i}^{n} \frac{p_i (T_i - Mean_j)^4}{Variance_j^2}
\tag{4}
]

where $p_i$ is the relative abundance and $T_i$ the trait value of the species \emph{i}, \emph{n} is the number of species in a community \emph{j} with available trait information and the sum of relative abundance is equal to 100\% for each community.
}
\examples{
result <- trait_moments(example_communitys, example_traits, n_species = 5, abundance = 80)
}
\references{
Le Bagousse-Pinguet, Y., Gross, N., Maestre, F.T., Maire, V., Bello, F. de & Fonseca, C.R. et al. (2017) Testing the environmental filtering concept in global drylands. Journal of Ecology, 105(4), 1058–1069. Available from: https://doi.org/10.1111/1365-2745.12735.
}
