% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_moments.R
\name{visualise_moments}
\alias{visualise_moments}
\title{visualise_moments}
\usage{
visualise_moments(result, traitID, comID, smooth = 1e+07)
}
\arguments{
\item{result}{The result of a calculation of trait distribution moments as returned by trait_moments.}

\item{traitID}{The ID of a trait.}

\item{comID}{The ID of one or several communities as vector.}

\item{smooth}{Degree of smoothness of the curves. Default is 10000000. Increasing the value improves the smoothness but also increases the calculation time.}
}
\value{
Visualization of trait distributions based on ggplot2.
}
\description{
visualise_moments visualises trait distributions for selected traits and communities based on density estimation from calculated moments.
See detailed information on in the github documentation
\href{https://github.com/SchreinerFR/TraitMoments}{TraitMoments}.
}
\examples{
\dontrun{
my_moment_plot <- visualise_moments(result = result,
traitID = "Trait.1",
comID = c(" Community.1", " Community.5", " Community.9"))

my_moment_plot

my_moment_plot +
 ggtitle("Distribution of trait 1 for three communities") +
 xlab("Trait 1") +
 scale_color_manual(values = c("black", "red", "orange")) +
 coord_cartesian(xlim = c(0,1), ylim = c(0,12), expand = FALSE)
}
}
