% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_moments.R
\name{trait_moments}
\alias{trait_moments}
\title{trait_moments}
\usage{
trait_moments(communities, traits, n_species = 4, abundance = 80)
}
\arguments{
\item{communities}{The species-plot-matrix as data frame. Column names must be species IDs and row names community IDs. It is not mandatory to provide relative abundances, but abundances should be on a nominal scale.}

\item{traits}{The species-trait-matrix as data frame. Column names must be trait IDs and row names species IDs. The species IDs must be consistent with those from the species-plot-matrix. Trait values have to be numeric.}

\item{n_species}{Number of predominant species for which trait information must be present. If the criterion is not met, NA is returned for the trait-plot-combination. Default is 4.}

\item{abundance}{Cumulative relative abundance for which trait information must be present. If the criterion is not met, NA is returned for the trait-plot-combination. Default is 80\%.}
}
\value{
Mean (CWM), variance, skewness and kurtosis of all trait distributions for each trait and each community.
}
\description{
trait_moments calculates all moments (mean, variance, skewness and kurtosis) of trait distribution across large community and traits datasets. The calculation is performed according to the equations number 1 to 4 from Le Bagousse-Pinguet et al. (2017).
See detailed information on the calculation in the github documentation:
\href{https://github.com/SchreinerFR/TraitMoments}{SchreinerFR/TraitMoments}
}
\examples{
result <- trait_moments(example_communitys, example_traits, n_species = 5, abundance = 80)
}
\references{
Le Bagousse-Pinguet, Y., Gross, N., Maestre, F.T., Maire, V., Bello, F. de & Fonseca, C.R. et al. (2017) Testing the environmental filtering concept in global drylands. Journal of Ecology, 105(4), 1058–1069. Available from: https://doi.org/10.1111/1365-2745.12735.
}
