% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_moments.R
\name{trait_moments}
\alias{trait_moments}
\title{trait_moments}
\usage{
trait_moments(communities, traits, n_species = 4, abundance = 80)
}
\arguments{
\item{communities}{The species-plot-matrix as data frame. Column names must be species IDs and row names community IDs.}

\item{traits}{The species-trait-matrix as data frame. Column names must be trait IDs and row names species IDs. The species IDs must be consistent with those from the species-plot-matrix.}

\item{n_species}{Number of predominant species for which trait information must be present. If the criterion is not met, NA is returned for the trait-plot-combination. Default is 4.}

\item{abundance}{Cumulative relative abundance for which trait information must be present. If the criterion is not met, NA is returned for the trait-plot-combination. Default is 80\%.}
}
\value{
Community weighted means and all higher moments (variance, skewness and kurtosis) of trait distributions.
}
\description{
trait_moments returns the community weighted means and all higher moments (variance, skewness and kurtosis) of trait distributions calculated from a species-plot-matrix and a species-trait-matrix.
}
\examples{
result <- trait_moments(example_communitys, example_traits, n_species = 5, abundance = 80)
}
